Tests passing
Implemented sqlite query for the "search" class method of Article.

Note that initial creation of Article with
`bin/rails generate model Article title:string body:text`

created a test/fixtures/articles.yml file that stored 2 article entries
-- this was messing up the first test, which checked that the Article
table was empty. I deleted the contents of this .yml but this is a good
note to have for later.