Get project up and running
Day 1 hour 0 of the take home. I roughly understand the overall structure of how things interact with each other. I have a server up and running and I understand how to display content. Next step is to figure out how this Article class should work in order for the unit tests to pass