Implement update functionality and partials
If I were to develop in ruby, I feel like I would need to read up on the docs to be familar with the tools required. I was able to transition into c++ and python quite easily (from Java) since a lot of the concepts had a very clean crossover. Ruby is quite a bit different however