Updated date showing
Tried to develop a more elegant form of date, by turning it into a
virtual variable that references updated_at. I did this by adding

def date
  updated_at
end

to app/models/article.rb

However, this did not comply with the vision of the tests -- date is an
actual field that can be modified.

So I went back and just edited the show.html.erb file to display
udpated_at if date is not set.